Files and Loops
This is not as beautiful as I'd like it to be but to be honest these topics together were abit challenging at first. 
Any way the input file is the .txt file i have attached, meant to be used as if the user specified a file in the same directory path as the .cpp  code executable file resides.